+++ src/valua/manifest.lua
@@ -41,6 +41,7 @@ return {
         parse = "valua.methods.parse",
         safe_parse = "valua.methods.safe_parse",
         is = "valua.methods.is",
+        alias = "valua.methods.alias",
         assume = "valua.methods.assume",
     },
 }