+++ src/valua/tooling/luals/analyzer/parser.lua
@@ -206,6 +206,19 @@ function parser.parse_tokens(tokens)
             else
                 idx = idx + 1
             end
+        elseif t and t.type == "identifier" then
+            local stmt_pos = t.pos
+            local start_idx = idx
+            local expr = parse_expr()
+            if expr and expr.type == "call" and expr.func == "alias" then
+                table.insert(declarations, {
+                    kind = "alias_statement",
+                    expr = expr,
+                    pos = stmt_pos,
+                })
+            elseif idx == start_idx then
+                idx = idx + 1
+            end
         else
             idx = idx + 1
         end